Change tab of some fields
