+++ frontend/src/components/manage/Blocks/ProjectsListing/Enterprises/constants.ts
@@ -67,22 +67,22 @@ export const enterpriseFieldsMapping: { [key: string]: string } = {
 export const initialOverviewFields = {
   name: '',
   country: null,
+  agency: null,
   location: '',
   city: '',
-  stage: '',
+  project_type: null,
   sector: null,
   subsector: null,
-  application: '',
-  local_ownership: null,
-  export_to_non_a5: null,
-  revision: null,
-  date_of_revision: null,
   status: null,
 }
 
 export const initialDetailsFields = {
-  agency: null,
-  project_type: null,
+  local_ownership: null,
+  export_to_non_a5: null,
+  application: '',
+  stage: '',
+  revision: null,
+  date_of_revision: null,
   planned_completion_date: null,
   actual_completion_date: null,
   project_duration: null,
