+++ frontend/src/components/manage/Blocks/ProjectsListing/Enterprises/edit/EnterpriseEdit.tsx
@@ -34,21 +34,21 @@ const EnterpriseEdit = ({ enterprise }: { enterprise: EnterpriseType }) => {
       overview: {
         name: enterprise.name,
         country: enterprise.country,
+        agency: enterprise.agency,
         location: enterprise.location,
         city: enterprise.city,
-        stage: enterprise.stage,
+        project_type: enterprise.project_type,
         sector: enterprise.sector,
         subsector: enterprise.subsector,
-        application: enterprise.application,
-        local_ownership: getFormattedDecimalValue(enterprise.local_ownership),
-        export_to_non_a5: getFormattedDecimalValue(enterprise.export_to_non_a5),
-        revision: enterprise.revision,
-        date_of_revision: enterprise.date_of_revision,
         status: enterprise.status,
       },
       details: {
-        agency: enterprise.agency,
-        project_type: enterprise.project_type,
+        local_ownership: getFormattedDecimalValue(enterprise.local_ownership),
+        export_to_non_a5: getFormattedDecimalValue(enterprise.export_to_non_a5),
+        application: enterprise.application,
+        stage: enterprise.stage,
+        revision: enterprise.revision,
+        date_of_revision: enterprise.date_of_revision,
         planned_completion_date: enterprise.planned_completion_date,
         actual_completion_date: enterprise.actual_completion_date,
         project_duration: enterprise.project_duration,
