+++ frontend/src/components/manage/Blocks/ProjectsListing/Enterprises/formTabs/EnterpriseOverviewSection.tsx
@@ -4,18 +4,11 @@ import EnterprisesDataContext from '@ors/contexts/Enterprises/EnterprisesDataCon
 import ProjectsDataContext from '@ors/contexts/Projects/ProjectsDataContext'
 import {
   EnterpriseTextField,
-  EnterpriseNumberField,
   EnterpriseSelectField,
-  EnterpriseDateField,
 } from '../FormHelperComponents'
 import useGetEnterpriseFieldsOpts from '../../hooks/useGetEnterpriseFieldsOpts'
 import { EnterpriseDataProps } from '../../interfaces'
-import {
-  textFields,
-  integerFields,
-  decimalFields,
-  dateFields,
-} from '../constants'
+import { textFields } from '../constants'
 
 import { map } from 'lodash'
 
@@ -35,7 +28,7 @@ const EnterpriseOverviewSection = ({
     mode,
   )
 
-  const { countries } = useContext(ProjectsDataContext)
+  const { countries, agencies, project_types } = useContext(ProjectsDataContext)
   const { statuses } = useContext(EnterprisesDataContext)
 
   const selectFields = [
@@ -44,6 +37,8 @@ const EnterpriseOverviewSection = ({
       options: countries,
       isDisabled: mode === 'edit',
     },
+    { fieldName: 'agency', options: agencies },
+    { fieldName: 'project_type', options: project_types },
     { fieldName: 'sector', options: sectors },
     { fieldName: 'subsector', options: subsectors },
     { fieldName: 'status', options: statuses },
@@ -55,11 +50,15 @@ const EnterpriseOverviewSection = ({
         field={textFields[0]}
         {...{ sectionIdentifier, ...rest }}
       />
-      <EnterpriseSelectField
-        field={selectFields[0]}
-        isDisabled={selectFields[0].isDisabled}
-        {...{ sectionIdentifier, ...rest }}
-      />
+      <div className="flex flex-wrap gap-x-20 gap-y-2">
+        {map(selectFields.slice(0, 2), (field, index) => (
+          <EnterpriseSelectField
+            key={index}
+            isDisabled={field.isDisabled}
+            {...{ sectionIdentifier, field, ...rest }}
+          />
+        ))}
+      </div>
       {map(textFields.slice(1, 3), (field, index) => (
         <EnterpriseTextField
           key={index}
@@ -67,39 +66,15 @@ const EnterpriseOverviewSection = ({
         />
       ))}
       <div className="flex flex-wrap gap-x-20 gap-y-2">
-        {map(selectFields.slice(1, 3), (field, index) => (
+        {map(selectFields.slice(2, 5), (field, index) => (
           <EnterpriseSelectField
             key={index}
             {...{ sectionIdentifier, field, ...rest }}
           />
         ))}
       </div>
-      <EnterpriseTextField
-        field={textFields[3]}
-        {...{ sectionIdentifier, ...rest }}
-      />
-      <div className="flex flex-wrap gap-x-20 gap-y-2">
-        {map(decimalFields, (field, index) => (
-          <EnterpriseNumberField
-            key={index}
-            dataType="decimal"
-            {...{ sectionIdentifier, field, ...rest }}
-          />
-        ))}
-      </div>
-      <div className="flex flex-wrap gap-x-20 gap-y-2">
-        <EnterpriseNumberField
-          dataType="integer"
-          field={integerFields[0]}
-          {...{ sectionIdentifier, ...rest }}
-        />
-        <EnterpriseDateField
-          field={dateFields[0]}
-          {...{ sectionIdentifier, ...rest }}
-        />
-      </div>
       <EnterpriseSelectField
-        field={selectFields[3]}
+        field={selectFields[5]}
         {...{ sectionIdentifier, ...rest }}
       />
     </div>
