+++ frontend/src/components/manage/Blocks/ProjectsListing/Enterprises/viewTabs/EnterpriseDetailsSection.tsx
@@ -1,7 +1,4 @@
-import { useContext } from 'react'
-
 import { getMeetingNr } from '@ors/components/manage/Utils/utilFunctions'
-import ProjectsDataContext from '@ors/contexts/Projects/ProjectsDataContext'
 import {
   dateDetailItem,
   detailItem,
@@ -10,7 +7,6 @@ import {
 import { detailsDateFields, enterpriseFieldsMapping } from '../constants'
 import { viewColumnsClassName } from '../../constants'
 import { EnterpriseType } from '../../interfaces'
-import { getEntityById } from '../utils'
 
 import { map } from 'lodash'
 
@@ -18,61 +14,72 @@ const EnterpriseDetailsSection = ({
   enterprise,
 }: {
   enterprise: EnterpriseType
-}) => {
-  const { agencies, project_types } = useContext(ProjectsDataContext)
-
-  const agency = getEntityById(agencies, enterprise.agency)?.name
-  const project_type = getEntityById(
-    project_types,
-    enterprise.project_type,
-  )?.name
-
-  return (
-    <div className="flex flex-col gap-4">
-      <div className={viewColumnsClassName}>
-        {detailItem(enterpriseFieldsMapping.agency, agency)}
-        {detailItem(enterpriseFieldsMapping.project_type, project_type)}
-      </div>
-      <div className={viewColumnsClassName}>
-        {map(detailsDateFields.slice(0, 2), (field, index) => (
-          <div key={index}>
-            {dateDetailItem(
-              enterpriseFieldsMapping[field],
-              enterprise[field as keyof typeof enterprise] as string,
-            )}
-          </div>
-        ))}
-        {numberDetailItem(
-          enterpriseFieldsMapping.project_duration,
-          enterprise.project_duration as string,
-          'integer',
-        )}
-      </div>
-      <div className={viewColumnsClassName}>
-        {detailItem(
-          enterpriseFieldsMapping.meeting,
-          getMeetingNr(enterprise?.meeting ?? undefined)?.toString() as string,
-        )}
-        {map(detailsDateFields.slice(2), (field, index) => (
-          <div key={index}>
-            {dateDetailItem(
-              enterpriseFieldsMapping[field],
-              enterprise[field as keyof typeof enterprise] as string,
-            )}
-          </div>
-        ))}
-      </div>
-      <div className="max-w-[90%]">
-        {detailItem(
-          enterpriseFieldsMapping.excom_provision,
-          enterprise.excom_provision,
-          {
-            detailClassname: 'self-start',
-          },
-        )}
-      </div>
+}) => (
+  <div className="flex flex-col gap-4">
+    <div className={viewColumnsClassName}>
+      {numberDetailItem(
+        enterpriseFieldsMapping.local_ownership,
+        enterprise.local_ownership as string,
+        'decimal',
+      )}
+      {numberDetailItem(
+        enterpriseFieldsMapping.export_to_non_a5,
+        enterprise.export_to_non_a5 as string,
+        'decimal',
+      )}
     </div>
-  )
-}
+    {detailItem(enterpriseFieldsMapping.application, enterprise.application)}
+    <div className={viewColumnsClassName}>
+      {detailItem(enterpriseFieldsMapping.stage, enterprise.stage)}
+      {numberDetailItem(
+        enterpriseFieldsMapping.revision,
+        enterprise.revision as string,
+        'integer',
+      )}
+      {dateDetailItem(
+        enterpriseFieldsMapping.date_of_revision,
+        enterprise.date_of_revision as string,
+      )}
+    </div>
+    <div className={viewColumnsClassName}>
+      {map(detailsDateFields.slice(0, 2), (field, index) => (
+        <div key={index}>
+          {dateDetailItem(
+            enterpriseFieldsMapping[field],
+            enterprise[field as keyof typeof enterprise] as string,
+          )}
+        </div>
+      ))}
+      {numberDetailItem(
+        enterpriseFieldsMapping.project_duration,
+        enterprise.project_duration as string,
+        'integer',
+      )}
+    </div>
+    <div className={viewColumnsClassName}>
+      {detailItem(
+        enterpriseFieldsMapping.meeting,
+        getMeetingNr(enterprise?.meeting ?? undefined)?.toString() as string,
+      )}
+      {map(detailsDateFields.slice(2), (field, index) => (
+        <div key={index}>
+          {dateDetailItem(
+            enterpriseFieldsMapping[field],
+            enterprise[field as keyof typeof enterprise] as string,
+          )}
+        </div>
+      ))}
+    </div>
+    <div className="max-w-[90%]">
+      {detailItem(
+        enterpriseFieldsMapping.excom_provision,
+        enterprise.excom_provision,
+        {
+          detailClassname: 'self-start',
+        },
+      )}
+    </div>
+  </div>
+)
 
 export default EnterpriseDetailsSection
