+++ frontend/src/components/manage/Blocks/ProjectsListing/Enterprises/viewTabs/EnterpriseOverviewSection.tsx
@@ -2,11 +2,7 @@ import { useContext } from 'react'
 
 import EnterprisesDataContext from '@ors/contexts/Enterprises/EnterprisesDataContext'
 import ProjectsDataContext from '@ors/contexts/Projects/ProjectsDataContext'
-import {
-  detailItem,
-  numberDetailItem,
-  dateDetailItem,
-} from '../../ProjectView/ViewHelperComponents'
+import { detailItem } from '../../ProjectView/ViewHelperComponents'
 import { enterpriseFieldsMapping } from '../constants'
 import { viewColumnsClassName } from '../../constants'
 import { EnterpriseType } from '../../interfaces'
@@ -18,57 +14,37 @@ const EnterpriseOverviewSection = ({
 }: {
   enterprise: EnterpriseType
 }) => {
-  const { countries, sectors, subsectors } = useContext(ProjectsDataContext)
+  const { countries, agencies, project_types, sectors, subsectors } =
+    useContext(ProjectsDataContext)
   const { statuses } = useContext(EnterprisesDataContext)
 
   const country = getEntityById(countries, enterprise.country)?.name
+  const agency = getEntityById(agencies, enterprise.agency)?.name
+  const project_type = getEntityById(
+    project_types,
+    enterprise.project_type,
+  )?.name
   const sector = getEntityById(sectors, enterprise.sector)?.name
   const subsector = getEntityById(subsectors, enterprise.subsector)?.name
   const status =
     find(statuses, (status) => status.id === enterprise.status)?.name ?? ''
 
   return (
-    <>
-      <div className="flex flex-col gap-4">
-        {detailItem(enterpriseFieldsMapping.name, enterprise.name)}
+    <div className="flex flex-col gap-4">
+      {detailItem(enterpriseFieldsMapping.name, enterprise.name)}
+      <div className={viewColumnsClassName}>
         {detailItem(enterpriseFieldsMapping.country, country)}
-        {detailItem(enterpriseFieldsMapping.location, enterprise.location)}
-        {detailItem(enterpriseFieldsMapping.city, enterprise.city)}
-        {detailItem(enterpriseFieldsMapping.stage, enterprise.stage)}
-        <div className={viewColumnsClassName}>
-          {detailItem(enterpriseFieldsMapping.sector, sector)}
-          {detailItem(enterpriseFieldsMapping.subsector, subsector)}
-        </div>
-        {detailItem(
-          enterpriseFieldsMapping.application,
-          enterprise.application,
-        )}
-        <div className={viewColumnsClassName}>
-          {numberDetailItem(
-            enterpriseFieldsMapping.local_ownership,
-            enterprise.local_ownership as string,
-            'decimal',
-          )}
-          {numberDetailItem(
-            enterpriseFieldsMapping.export_to_non_a5,
-            enterprise.export_to_non_a5 as string,
-            'decimal',
-          )}
-        </div>
-        <div className={viewColumnsClassName}>
-          {numberDetailItem(
-            enterpriseFieldsMapping.revision,
-            enterprise.revision as string,
-            'integer',
-          )}
-          {dateDetailItem(
-            enterpriseFieldsMapping.date_of_revision,
-            enterprise.date_of_revision as string,
-          )}
-        </div>
-        {detailItem(enterpriseFieldsMapping.status, status)}
+        {detailItem(enterpriseFieldsMapping.agency, agency)}
       </div>
-    </>
+      {detailItem(enterpriseFieldsMapping.location, enterprise.location)}
+      {detailItem(enterpriseFieldsMapping.city, enterprise.city)}
+      <div className={viewColumnsClassName}>
+        {detailItem(enterpriseFieldsMapping.project_type, project_type)}
+        {detailItem(enterpriseFieldsMapping.sector, sector)}
+        {detailItem(enterpriseFieldsMapping.subsector, subsector)}
+      </div>
+      {detailItem(enterpriseFieldsMapping.status, status)}
+    </div>
   )
 }
